requirements.txt
BTrees==4.7.2
cffi==1.14.0
click==7.1.2
contextlib2==0.6.0.post1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
persistent==4.6.4
pycparser==2.20
python-dateutil==2.8.1
schema==0.7.2
six==1.15.0
transaction==3.0.0
Werkzeug==1.0.1
zc.lockfile==2.0
ZConfig==3.5.0
ZODB==5.6.0
zodbpickle==2.0.0
zope.interface==5.1.0
